+++ packages/manager/apps/hycu/src/constants.ts
@@ -7,3 +7,8 @@ export const packTypeLabel = {
   'hycu-cloud-vm-pack-250': '250 VMs',
   'hycu-cloud-vm-pack-500': '500 VMs',
 } as const;
+
+export const HYCU_CHANGE_PACK_FEATURE_ACTIVATED = false;
+
+export const LICENSE_FILE_EXT = '.req';
+export const LICENSE_FILE_NAME_TEMPLATE = `license-hycu-{serviceName}.dat`;