+++ packages/manager/apps/hub/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.0.3](https://github.com/ovh/manager/compare/@ovh-ux/manager-hub-app@6.0.2...@ovh-ux/manager-hub-app@6.0.3) (2024-11-20)
+
+**Note:** Version bump only for package @ovh-ux/manager-hub-app
+
+
+
+
+
 ## [6.0.2](https://github.com/ovh/manager/compare/@ovh-ux/manager-hub-app@6.0.1...@ovh-ux/manager-hub-app@6.0.2) (2024-11-19)
 
 **Note:** Version bump only for package @ovh-ux/manager-hub-app